refactor: improve mappings classes readability
