+++ src/Attribute/BelongsToOrganisations.php
@@ -2,22 +2,22 @@
 
 declare(strict_types=1);
 
-namespace LaravelDoctrine\ACL\Mappings;
+namespace LaravelDoctrine\ACL\Attribute;
 
 use Attribute;
-use Illuminate\Contracts\Config\Repository;
+use Illuminate\Contracts\Config\Repository as Config;
 
 #[Attribute(Attribute::TARGET_PROPERTY)]
-final class HasPermissions extends RelationAttribute
+final class BelongsToOrganisations extends RelationAttribute
 {
     public function __construct(
         string|null $targetEntity = null,
         string|null $mappedBy = null,
-        public string|null $inversedBy = null,
         array|null $cascade = null,
         string $fetch = 'LAZY',
         bool $orphanRemoval = false,
         string|null $indexBy = null,
+        public string $inversedBy = 'users',
     ) {
         $this->targetEntity  = $targetEntity;
         $this->mappedBy      = $mappedBy;
@@ -27,13 +27,8 @@ final class HasPermissions extends RelationAttribute
         $this->indexBy       = $indexBy;
     }
 
-    public function getTargetEntity(Repository $config): string|null
+    public function getTargetEntity(Config $config): string|null
     {
-        // Config driver has no target entity
-        if ($config->get('acl.permissions.driver', 'config') === 'config') {
-            return null;
-        }
-
-        return $this->targetEntity ?: $config->get('acl.permissions.entity', 'Permission');
+        return $this->targetEntity ?: $config->get('acl.organisations.entity', 'Organisation');
     }
 }
