+++ src/Attribute/MappingAttribute.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace LaravelDoctrine\ACL\Mappings;
+namespace LaravelDoctrine\ACL\Attribute;
 
 use Doctrine\ORM\Mapping\MappingAttribute as DoctrineMappingAttribute;
-use Illuminate\Contracts\Config\Repository;
+use Illuminate\Contracts\Config\Repository as Config;
 
 interface MappingAttribute extends DoctrineMappingAttribute
 {
-    public function getTargetEntity(Repository $config): string|null;
+    public function getTargetEntity(Config $config): string|null;
 }
