+++ src/Mappings/Builders/Builder.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace LaravelDoctrine\ACL\Mappings\Builders;
 
 use Doctrine\Persistence\Mapping\ClassMetadata;
-use LaravelDoctrine\ACL\Mappings\MappingAttribute;
+use LaravelDoctrine\ACL\Attribute\MappingAttribute;
 use ReflectionProperty;
 
 interface Builder
