+++ workbench/app/Entities/UserSingleOrg.php
@@ -12,10 +12,10 @@ use Illuminate\Contracts\Auth\Access\Authorizable as AuthorizableContract;
 use Illuminate\Contracts\Auth\Authenticatable as AuthenticatableContract;
 use Illuminate\Contracts\Auth\CanResetPassword as CanResetPasswordContract;
 use Illuminate\Foundation\Auth\Access\Authorizable;
+use LaravelDoctrine\ACL\Attribute\BelongsToOrganisation;
 use LaravelDoctrine\ACL\Contracts\BelongsToOrganisation as BelongsToOrganisationContract;
 use LaravelDoctrine\ACL\Contracts\HasPermissions as HasPermissionsContract;
 use LaravelDoctrine\ACL\Contracts\HasRoles as HasRolesContract;
-use LaravelDoctrine\ACL\Mappings as ACL;
 use LaravelDoctrine\ACL\Organisations\BelongsToOrganisation as TraitBelongsToOrganisation;
 use LaravelDoctrine\ACL\Permissions\HasPermissions;
 use LaravelDoctrine\ACL\Roles\HasRoles;
@@ -48,14 +48,14 @@ class UserSingleOrg implements AuthenticatableContract, AuthorizableContract, Ca
     public string $email;
 
     /** @var Collection<int, Role> */
-    #[ACL\HasRoles()]
+    #[\LaravelDoctrine\ACL\Attribute\HasRoles()]
     public Collection $roles;
 
     /** @var array<string> */
     #[ORM\Column(type: 'json')]
     public array $permissions = [];
 
-    #[ACL\BelongsToOrganisation()]
+    #[BelongsToOrganisation()]
     public Organisation|null $organisation = null;
 
     public function __construct()
